report-page-url-conflict
Super Admin rapor oluşturamıyor (departmanı yok)
dashboard sayfası ile staj arama sayfası yer değiştirdi.
öğrenci değerlendirme sayfasının extra özellikleri ekli değil

staj arama sayfası reponsive oldu.
pagination var.
rapor oluşturma sayfası responsive ve hazır sadece gerekli bilgilerle docx oluşacak.

öğrenci entitysine grade eklendi olmadığı için null dönüyordu bu da staj arama sayfasındaki filtrelemeyi etkiliyordu.
