+++ backend/src/routes/searchStudentRoutes.ts
@@ -1,0 +1,10 @@
+import express from 'express';
+import { verifyToken } from '../middlewares/verifyToken';
+import searchStudentController from '../controllers/searchStudentController';
+
+const router = express.Router();
+
+router.get('/internships', verifyToken, searchStudentController.getStudents);
+router.put('/internships/:id/state', verifyToken, searchStudentController.putInternshipState);
+
+export default router;
