--- front/src/pages/DashboardActual/DashboardActual.css
@@ -1,15 +1,0 @@
-.secondChartFilterLine {
-    gap: 10px;
-    width: 20em;
-    display: flex;
-    flex-wrap: nowrap
-}
-
-@media (max-width: 768px) {
-    .secondChartFilterLine {
-        gap: 10px;
-        width: 100%;
-        display: flex;
-        flex-wrap: wrap
-    }
-}
